Fix incorrect handling of stream capacity greater than 2^32 (#806)
Motivation:

We did cast the stream capacity to an jint while it should be jlong.
This could cause an erroneous closure of the stream in
QuicheQuicChannel.java:handleWritableStreams().

Modifications:

Correctly cast to jlong

Result:

No more erroneous closure caused by incorrect casting. Fixes
https://github.com/netty/netty-incubator-codec-quic/issues/805

+++ codec-classes-quic/src/main/java/io/netty/incubator/codec/quic/Quiche.java
@@ -416,7 +416,7 @@ final class Quiche {
      * See
      * <a href="https://github.com/cloudflare/quiche/blob/0.6.0/include/quiche.h#L274">quiche_conn_stream_capacity</a>.
      */
-    static native int quiche_conn_stream_capacity(long connAddr, long streamId);
+    static native long quiche_conn_stream_capacity(long connAddr, long streamId);
 
     /**
      * See
